Add clang version string to clang_delta --version
This adds the version of clang currently in use to the output from
clang_delta --version and changes the test script to use this information
instead of looking at the list of dynamic libraries when checking which clang
version is in use. This allows the tests to run when clang_delta is statically
linked against libClang.