Fix hang when buggy (unaltered) OK comes after STOP (#162)
Make sure that we don't proceed with an unmodified file - for this,
consistently validate the pass' results on both codepaths (in
process_done_futures() and wait_for_first_success()).

Besides fixing the hang that was possible when a pass misbehaved,
there are a couple of other side effects:

* more cases of ERRORs in a pass will be reported - previously only
  the ones on the first codepath were handled;
* max_improvement is enforced stricter now - previously the second
  codepath allowed a larger reduction to sneak in.

This fixes #156.