hints: Introduce basic hint (text deletion) format & helpers (#175)
Hints are the proposed format for being used by reduction heuristics. A hint describes one or several modifications (e.g., deletion of text at particular location) to be attempted on the test case during reduction.

This uniform intermediate representation will bring the following benefits, both for performance as well as versatility:

* It's easier to implement reduction heuristics - they don't need to take care of editing the text, implementing binary search for the sake of performance, etc.

* Allow different heuristics to be attempted concurrently, in an interleaving fashion.

* Successful reductions that were concurrently discovered can be folded (merged) instead of proceeding with only one reduction.

This commit lays the groundwork for #169 - the format (JSON Schemas) and implementation of the most primitive hint format: simple text deletion.