hints: Migrate lines pass
This reimplements the Lines pass using the hint-based architecture.

Besides long-term improvements brought by the hint approach, this brings
an actual functional improvement for this heuristic: no reformatting of
input files is needed. Instead of overwriting the file in the "one block at
a particular depth per line" fashion and deleting lines in this new text,
we just emit hints for deletable code blocks.

This implies the following practical benefits:

1. The heuristic won't degrade dramatically when reformatting leads to a
   broken code (e.g., observed for C++14/C23 digit separators in #166);
2. Whitespaces aren't added/removed at various places in this pass;
3. Check_sanity is not needed anymore - we don't need to spend time
   checking whether the reformatted input still satisfies the
   interestingness test;
4. (minor) In C-Vise logs, the number of lines won't increase/decrease
   unpredictably whenever a Lines pass with some argument is started.

This also unblocks implementing the following improvements in the future:

5. Concurrently discovered reductions can be folded (merged) easily -
   there are no unrelated formatting changes that'd make merging very
   hard.
6. The Lines pass can run concurrently with other passes in an
   interleaving fashion, without having to maintain an extra input copy.