Improve Lines: only remove exactly at given depth (#192)
Stop producing hints for deleting code at levels smaller than the specified
depth (threshold). This should significantly improve the success rate for
all depths greater than zero, because we'll stop trying partially removing
outer constructs.

For example, for a top-level C function, running Lines with arg=1 should
only attempt removing statements inside the function body, but don't
try deleting the function signature alone or deleting the closing curly
brace alone.

Note: the downside of this change is that when running Lines for arg>1,
the heuristic will stop trying to delete higher-level chunks of code. We
consider this not to be a problem because one of planned follow-up
changes will run all Lines passes concurrently, in an interleaving fashion.