interleaving: Manager can run different passes simultaneously (#199)
Add basic support of executing different passes in an interleaving fashion to TestManager.

This paves the way for improving performance, both worst-case (to avoid C-Vise being
stuck on a badly performing heuristic meanwhile others would reduce the file quicker)
as well as average-case (once we implement folding of concurrently-discovered
reductions). We only plan to enable the interleaving mode for hint-based passes,
because they are designed to leave inputs unmodified and to have easily mergeable
reductions.

This commit doesn't wire up new logic to the CLI; only tests exercise this mode at
the moment.