hints: Add CLI for applying hints to input (#201)
Add command-line options to C-Vise to apply the given set of hints
to the test case.

This is currently intended to be used in clang_delta unit tests
(which, despite being written in Python, are not part of the C-Vise
module and hence cannot import hints.py directly). Additionally,
this might be useful in the long run for developers working on new
passes for C-Vise.