cli: Don't close stderr (#211)
The "with sys.stderr" construct was closing the stderr file
descriptor. While this happened near the program exit, it was still
error-prone since if there are background threads or processes
they'd start hitting (highly unusual) errors when printing to
stderr.

In particular, this closed stderr would've broken the typical
"Logging to a single file from multiple processes" pattern (which
we're planning to switch to in order to switch to the forkserver
multiprocessing mode).