hints: Support text replacement patches (#209)
Implement support for hints that replace text with the new value.
In order to have a compact representation, use a "vocabulary" - a
string table that hints can refer to; the vocabulary can also be
used for other features in the future.

This is preparation for porting heuristics that do text replacements
to the hint-based framework.