lines: Don't generate hints for whitespace-only blocks (#223)
Enhance the topformflat Flex grammar rules to only generate
nontrivial hints, i.e. those consisting of non-whitespaces.
Empirically, this reduces the number of hints by around 20%;
and extra whitespaces are easy to remove at the end of reduction.

As a byproduct of this change, we were able to stricten some
unit tests - now the distracting whitespace-only removals don't
happen.

Another functional change: only treat "#" characters as a start
of a macro if it's the first character or there are only whitespaces
before it; this should be a pure improvement too.