hints: Migrate comments pass (#227)
Similar to other commits in the series, porting a pass to the hint-based
architecture allows for better parallelization in the future.

I've ported the heuristic as-is - even though it being a simple regexp
doesn't handle many cases correctly, like `foo("http://example.com")`
being mistreated as a comment because of the double-slash.