hints: Migrate clex::rm-toks pass
Port the "rm-toks-N" command from the "clex" pass - the one that
removes N consecutive tokens based on a (simplistic) C grammar. The
logic remains the same as before - the pass is just switched to
using hints.

Here we also introduce a new parameter into the hint bundle -
"grouping", so that we can disable the binary search logic for this
pass "clex::rm-toks-N" (it'd be wasteful - there are other passes
that are much better at recognizing big removable chunks of input).