test_manager: Tolerate non-UTF-8 in interestingness test
Don't abort with an exception when the interestingness test prints
non-UTF-8 data - it's not an error since only the exit code matters.
We only need to do something when the logging is enabled (--debug
and the bad interestingness test), in which case we filter out bad
bytes.

This change should also bring some (small) performance improvement,
since we stop doing UTF-8 decoding unilaterally on all child
processes outputs.