test_manager: Don't override Pebble sleep unit
This was documented as a measure to increase the responsiveness,
but according to my measurements it only slows down C-Vise: e.g.,
simple reductions finish in ~113 sec instead of ~86 sec (30%
slowdown).

Presumably the slowdown is caused by the increased thread
congestion, which with Python's Global Interpreter Lock decreases
the main process' throughput.