Override child processes' temp dir (#262)
When executing an interestingness test, make it use a custom temp dir,
located inside the job's temporary folder, instead of the default location
(like /tmp/ on a typical POSIX system). This allows C-Vise to clean up all
temporary files it might've created without cluttering the default
location.

Examples of problematic cases which this commit attempts to solve:

* The interestingness script launches GCC which creates a
   temporary `.s` file; meanwhile it's normally removed on GCC
   process exit, it doesn't happen when C-Vise kills cancelled jobs
   via SIGKILL.
* The interestingness script launches Clang without
   `-fno-crash-diagnostics`; once Clang crashes, it writes a the crash
   reproducer file to the temporary directory.

This fixes #260 (hopefully).