interleaving: Concurrently execute advance_on_success (#261)
Each time we commit to a reduction, schedule all passes'
advance_on_success() methods to be executed in workers in parallel,
as opposed to running them sequentially.

This significantly improves the performance of the "interleaving" mode,
since we previously were spending 20-40% of wall clock time in those
single-job advance_on_success() calls for heavy inputs.

Example measurements on a 64-core machine for the "gcc-94937"
reproducer:
* before: reduced 54.4% after 5 min, 78.7% after 10 min,
* after: reduced 58.1% after 5 min, 85.2% after 10 min.