test_manager: Rewind to earliest unfinished on committing
When committing to a reduction candidate, rewind each pass state
back to the earliest state which testing hasn't finished yet.

For example, if we scheduled 4 jobs and at some moment the status
is "job1=failed, job2=running, job3=succeeded, job4=running" and
we want to proceed with job3's successful transformation, next
time we want to recheck job2.

The main benefit is middle/late stages of the reduction when it's
important to test every single reduction possibility without
skipping ant of them.