tree-sitter-cpp: Import 0.23.4 (#269)
Add a new third-party dependency: tree-sitter-cpp - C++ grammar for
tree-sitter. We're going to use it, in conjunction with Tree-sitter,
to implement new reduction heuristics.

Note: Similarly to the Tree-sitter library, we import this grammar
directly as opposed to using system-wide packages (largely nonexisting
for tree-sitter-cpp) or git submodules (hard to maintain).