hints: Migrate clex::rm-toks pass (#263)
Port the "rm-toks-N" command from the "clex" pass - the one that
removes N consecutive tokens based on a (simplistic) C grammar.

The main logic remains the same as before, except that we group
the enumeration in a few concurrently executed groups: "try
ranges of 1 to 4 consecutive tokens", "5 to 8 tokens", etc.
Empirically this seems to work better than putting all "1 to 16" or
"16 to 1" token ranges into the same pass, because we can
benefit from few-token removals in parallel to going for bigger
reductions. The specific constants might need to be tweaked
further in the future.

We also add randomization to the enumeration order (simply
starting from a random shift - a completely random permutation
seems hard to achieve due to implementation/performance
concerns).

Another change we make is treating a trailing backslash in a line
as a separate token. Without this, these backslashes sometimes
remained undeleted, even on the simple test blocksort-part.c
(this likely wasn't happening before due to a slightly different
order of pass application).