Stats show size improvement for each pass
Measure how many bytes were removed from the test case using each
individual pass. Sort the printed statistics by this criteria (instead
of the total time spent on a pass).

This gives a better understanding of which passes work well and which
don't - especially in the first half of the reduction, when inputs are
big and it's important to quickly prune big chunks. However, one
should be cautious because the new statistic doesn't give high score
to passes which are naturally small-reducers (like "rm-toks") despite
being very crucial for the end result.