hints: Use msgspec for faster [de]serialization (#272)
Use the msgspec library for faster JSON parsing and dumping. According
to our measurements, this speeds up hint saving by ~22x and hint
loading by ~1.5x. The overall effect on the reduction performance seems
to be around 2%.

Note that we also tried the orjson library, but the speed improvement
turned out to be substantially smaller there.