hints: Migrate Balanced pass (#277)
We reimplement the pass from scratch, since the previously used
"nestedmatcher" library is quite complicated and hard to combine
with the idea that all transforms should be pregenerated at once
and in form of patches (hints).

Also binary search is enabled for this pass - empirically this seems
to perform slightly faster than trying each transformation one-by-one.