tree-sitter: Add EraseNamespace pass (#280)
This deletes everything that's inside a namespace' curly braces - one attempt
(one hint) per namespace, with a binary search logic on top.

This pass' usefulness varies from test to test, but in general it seems like a
valuable addition to C-Vise. Average success rate: 15% (with the binary search),
average number of bytes reduced while working on a given test: 5%.