clang_delta: Skip non-dep typedefs in replace-dependent-typedef (#276)
Only consider typedefs that depend on other typedefs in
replace-dependent-typedef.

This prevents the pass from reporting a success when no actual change
has been made, like in the case when a typedef belongs to and refers to
a template struct (where multiple possible string representations may
confuse the checks).

This fixes the problem with this heuristic highlighted in #156.

Note: essentially the same change was originally proposed in the upstream -
C-Reduce - however that project seems unmaintained at the moment:
https://github.com/csmith-project/creduce/pull/281