interleaving: Improve fold success rate by banning failures (#287)
Further boost the benefits of the interleaving mode by a cleverer
folding strategy: when all candidates together don't pass the
interestingness test together, heuristically pick up a subset of
them that should be more likely to work.

The strategy is simple: take the largest known-to-succeed fold,
then subtract it from each known-to-fail fold, then mark a
randomly selected half of each as "banned". All remaining states
should have, roughly, 50%-75% chance to succeed when taken
together.

For example, suppose the input contains 4 functions f1, f2, f3, f4,
and some pass discovered that each of them can be removed
individually, and then fold "remove f1+f2" succeeded however
"remove f1+f2+f3" failed. Then the new logic in this commit will
infer that the next fold to be attempted should be "remove
f1+f2+f4" instead of the likely-to-fail "f1+f2+f3+f4".

The overall improvement on the C-Vise performance seems to
be between 0% and 30%, depending on the test case.