Refactor code to use pathlib for paths (#289)
Pathlib.Path is a more convenient and reliable way to work with
file paths. Change the C-Vise code base to use it consistently, as
opposed to using Str in half of places.

This is a pure refactoring change, but it'll be handy for the
follow-up changes that bring multi-file (directory) reduction.