Optimize input copying: offload to workers; skip for hint passes (#290)
Before this commit, before any reduction attempt is made
(Pass.transform()), the input test was copied, and this happened
in the main C-Vise process.

The commit optimizes this in two aspects:

1. Copying is moved to the worker processes. This improves throughput,
since copying is likely to be well parallelizable (at least for typical *nix
tmpfs setups), and it makes the main process' scheduler loop more
responsive.

2. Copying is skipped altogether for hint-based passes; they create the
working copy by applying hints to the original file. In theory, we could
change all passes to work in this way - but the implementation effort
would be higher, and only hint-based passes are executed in the
beginning, when the inputs are still huge.

This optimization seems to bring between 0 and 10% speedup to
C-Vise, depending on the test case. Additionally, it simplifies the
upcoming work on multi-file/directory reductions.