interleaving: Choose candidate by size (and other criteria) (#292)
When multiple successful discoveries are made, generally prefer the one
which reduces the test case more significantly - instead of the previous
hacky "the first/the last attempt always wins".

We only make two special cases to the new rule. First, some passes do
increase the test size as part of normal operation (e.g., the typedef
inlining); for such cases, the size criteria would have the deterrent
effect, hence when the size increases we decided to prefer to include as
much transformations as possible. Second, we prefer folds over
non-folds, since it's beneficial in the long run to build up a diverse, even
if a temporarily not the best, set of patches.

The performance impact on C-Vise reduction speed seems to vary
between 30% speedup and 10% slowdown; we assume the slowdowns
are occasional cases where we get "unlucky" by picking up a local
optimum that's not the globally optimal path towards the result. So
despite this, the change still makes the algorithm easier to reason
about, and the edge cases can be tackled in the future.