Terminate subprocess on exception; robuster PID queue (#303)
Terminate the child process (a pass helper tool, like "clex") when an
exception occurs - most importantly, when a KeyboardInterrupt or
SystemExit exceptions happen in the worker. Change more passes to
record the child process PID in the PID queue. Make the code more
exception-safe.

Both changes should improve the robustness of C-Vise and reduce the
number of cases when child processes "leak" and keep eating system
resources after the reduction job finished.