test_manager: Replace files atomically when committing (#304)
Instead of directly writing to the test case, write first to a temporary
file and then rename it. This is atomic on popular file systems. We also
add special code for KeyboardInterrupt/SystemExit to make it robust
to random terminations.

Also refactor the code to have a dedicated fileutil helper, and cover it
with unit tests.