test_cvise: Use a shorter TMPDIR (#308)
For the TMPDIR value that we pass when executing C-Vise in integration
tests, create a fresh temp dir in a regular way (e.g., under /tmp/ on a
typical *nix) instead of stacking one inside the pytest temp path.

This shorter TMPDIR value reduces the possibility of hitting the "path
too long" errors, especially when using pytest-xdist, which makes
pytest tmp_path longer.