Use forkserver multiprocessing mode (#305)
This fixes potential deadlocks caused by forking a multi-threaded program,
and follows generic Python guideline. Also it reduces the RAM pressure since
workers don't get copies of the main C-Vise process data structures, like the
pass cache (despite copy-on-write allowing to avoid physically copying those
pages on fork, programs like OOM monitors might treat this as real usage).

The downside is that it causes a 25-50% slowdown compared to the "fork" mode,
because forkserver workers are slower to start and interact with. Part of this
seems to be caused by https://github.com/python/cpython/issues/137996,
which will hopefully fixed in new Python versions. The rest of slowdown we'll
mitigate by further optimizations in C-Vise in follow-up commits.