subprocesses: Signals don't interrupt child spawn (#324)
Protect the code that creates a subprocess from signal handling, so
that we do always (*) report the started process' PID. If a signal
occurs within this time frame, the exception will be raised after this
mini-transaction ends.

(*) If the process creation is very slow and happens to collide with
the time point when the Pebble worker gets canceled, the PID would
remain unreported. We consider this a very rare case, and it can be
solved by using process groups in the future.