hotkeys: Fix garbled terminal on unclean exit; refactor
Fix the console sometimes remaining garbled (typed characters not visible)
when C-Vise exits.

The problem was that the KeyLogger code was changing the terminal's
properties, and unsetting them back in an object finalizer - however
finalizers' semantics is vague and it's possible that a new KeyLogger is
created before the previous one is finalized. This commit changes the
KeyLogger to be a context manager with a clear lifetime. Additionally we
refactor the code to split out Windows-specific code, for the sake of
readability.