Add timeouts for pass initialization jobs
Impose a timeout for pass initializations, so that a hung pass never hangs
the whole C-Vise process. Additionally, this change allows us to gracefully
handle the bug in Pebble when a scheduled task may occasionally be lost.

We set the init timeout to be a generous 10x of the regular --timeout value,
as a safe default without having the user to guesstimate the parameter
themselves.