multiprocessing: Workaround for hung Pebble jobs (#329)
Pebble seems to have a race condition, resulting in jobs occasionally "hanging":
the future never gets resolved with a result nor with a timeout. This results
in hanging the whole C-Vise process, since it waits for a completion of a job
infinitely long in case no previous jobs in the batch succeeded.

This commit adds two workarounds for this issue:

1. After we cancel jobs, run a special batch of N sleep no-op tasks - they
should bear the brunt. We detect which of those tasks get hung and cancel them,
while letting all other tasks complete. This should avoid the race condition
(a worker being canceled for taskA manages to start doing taskB) at the cost
of [0.1; 60] seconds penalty every time C-Vise commits to a reduction.

2. Introduce another timeout check: the maximum time between the task's schedule
time and the task completion. We let it be a generous (10x) multiplier of a
regular timeout. This is only intended as a backup solution if the
solution \#\1 doesn't help against the hang.