interleaving: Make fair round-robin pass scheduling
Try to schedule the same number of jobs for each pass, as opposed to
picking the next pass according to a cyclic counter.

This makes the scheduling fairer, without devoting exceedingly many jobs
to the pass that was the first to initialize (typically some simple pass,
like "blank") and without giving too few jobs to slow-to-init passes (like
clang_delta based ones).