cache: Store in temp files; use hashes; limit per LRU (#344)
Change the cache implementation to store contents as files (in temp dirs) as
opposed to keeping them all in the memory of the C-Vise process. Only keep a
few last items instead of caching forever. Use hashing via SHA-256 to perform
cache lookups instead of full file contents. Write to the test case atomically
when taking a value from cache.

This significantly reduces the memory usage by the C-Vise main process and makes
cache operations faster. This also prevents the test case file corruption if
Ctrl-C is pressed when a cached value is written. Additionally, this paves the
way for supporting multi-file (directory) test case caching.