Fix occasional log spam from signal handlers (#345)
This suppresses raising exceptions in some cases where they'll anyway be likely
discarded by the Python standard library: inside __del__() methods and inside
weakref finalizers.

This fixes the "Exception ignored in" logs occasionally appearing in the C-Vise
stderr, in case when Pebble was trying to terminate a worker that was in a
middle of one of such operations. This commit doesn't intend to be a
comprehensive solution, but rather a guard for relatively common cases.

Also fix the typo that caused KeyboardInterrupt being raised on SIGTERM
(instead of SIGINT), and cover sigmonitor with unit tests.