Observe signals in event loop via future instead of timeout (#347)
Get rid of unnecessary iterations and timeouts in the C-Vise main process event
loop, and use futures to immediately exit the loop when a signal is observed.

This removes some unnecessary work from the otherwise busy main loop, and makes
the process more responsive even signals occur at unexpected times. As a bonus,
we get rid of one magic constant in the code.