multifile: Create @fileref in ClangModuleMap (#350)
Extend the ClangModuleMap pass to additionally report @fileref hints for each
mentioned header file. To store the references, extend the hint data schema to
have a new field "extra".

This commit by itself doesn't affect the reduction process, however it's
preparation for the introduction of the unused file removal pass.