multifile: Assume all Makefiles to be used
Suppress removal of Makefiles in the RmUnusedFilesPass, because in a
typical use case the makekfile would be used by the interestingness
test to compile the source.

An ideal solution would be to detect whether a makefile is actually
used during the interestingness test's invocation, but currently C-Vise
cannot make any insights into the interestingness tests, and in
general the script there may be arbitrarily complicated.