Fix Lines to delete last unfinished block in a file
Improve the LinesPass to also emit a hint for the last chunk in the
file when it's unfinished. For example, for a file ending with
"namespace {" we'll now attempt deleting this ending for arg="0".
Such examples occur often in the middle of reduction, or when macros
with curly brackets confuse the nesting calculation.

Also make topformflat robust to file opening failures, similarly to
what was done to Clex rm-toks in #365.