ClangIncludeGraph: Support Clang Header Modules (#367)
Support Clang C/C++ header modules when extracting the include graph in ClangIncludeGraphPass.

Implementation-wise, we have to manipulate the command line because "-fmodule-file=..." arguments would refer to non-existing PCM files, so these and "-fmodule-map-file=..." arguments that depend on them have to be deleted.

On the other hand, "-fmodules" has to be kept, otherwise we won't be able to easily run Clang's preprocessor on a modulemap file. Also "-fmodule-map-file-home-is-cwd", if present, has to be kept - otherwise Clang won't find the headers mentioned in the modulemap.