ClangIncludeGraph: Parallelize pass initialization (#369)
As running clang_include_graph sequentially may take a lot of time
(several minutes) on big inputs with many compilation commands, we split this
work into K workers that are executed concurrently - this way we reduce the
wall clock time it takes for the pass to finish initialization. Currently the
hardcoded value of K is 10.

Implementation-wise, to avoid introducing new complexity due to new
parallelization mechanisms, we reuse the existing functionality: we spawn
K "subordinate" passes that do the actual work, and then the main pass just
merges the hints produced by them all.