clangincludegraph: Take makefile paths from MakefilePass (#392)
With this commit, ClangIncludeGraphPass obtains the list of makefiles from
MakefilePass (using the special @makefile hints) instead of finding the files
itself.

This allows us to remove the hardcoded list of makefile file patterns from the
code - the "claim_files" property of the "makefile" pass in the C-Vise JSON
config becomes the source of truth.