hints: Preserve empty dirs in the test case
This replicates empty directories from the input test case when creating
a working transformed copy of a multi-file input.

Before this commit, directories were only created when there's at least
one file in them, which prevented some reduction possibilities. While we
do want to be able to clean up empty directories, we should do this properly,
only for directories that aren't referenced by any command line - this will
be added in a follow-up.