rmunusedfiles: Removal of empty unused directories
Support the "rm" hint operation for directories, and implement automatic
removal of unused empty directories in the RmUnusedFilesPass.

Note that we try to avoid removing used directories, like those mentioned
in "-I" compilation parameters in makefiles, since that might result in the
reproducer accidentally using system headers. The solution in this commit
is still heuristic and not fully comprehensive, so if one really wants to
prevent system headers from being used they should check it in the
interestingness test.