multifile: Remove files created by interestingness test
Remove extraneous files/directories that the interestingness test might've
created in the test case directory (for multifile inputs).

This addresses the common use case when the interestingness test runs a
compiler in the test case directory, so that object files etc. ends up in
the folder. Previously, the interestingness test's author was responsible
for cleaning them up, and failing to do so would result in C-Vise copying
these files back to the test case, producing weird results and stats. Now
the removal will happen automatically.