multifile: Remove files created by interestingness test (#424)
Remove extraneous files/directories that the interestingness test might've
created in the test case directory (for multifile inputs).

This addresses the common use case when the interestingness test runs a compiler
in the test case directory, so that object files etc. ends up in the folder.
Previously, the interestingness test's author was responsible for cleaning them
up, and failing to do so would result in C-Vise copying these files back to the
test case, producing weird results and stats. Now the removal will happen
automatically.

Another, a bit trickier, case was if the interestingness test did "mkdir" for
output dirs. If no one cleans up them after the interestingness test, C-Vise
would essentially hang, since a pass would report that it deleted a directory,
which would then reappear in the test case and move us back to square one.