test_manager: Log initial file sizes on startup
This is slightly more informative since it gives an idea of the starting
point (zero procent but nonzero sizes).

Additionally, for multi-file test cases, this prodives the user with
specific numbers that would otherwise be not trivial to know (especially
the total size - this is typically smaller than the "du" output which adds
up file system blocks).