test_manager: Log initial file sizes on startup (#425)
This is slightly more informative since it gives an idea of the starting point
(zero percent but nonzero sizes).

Additionally, for multi-file test cases, this prodives the user with specific
numbers that would otherwise be not trivial to get (especially the total size -
this is typically smaller than the "du" output which adds up file system
blocks).