test_manager: Fix wait when missing timeout workaround is due  (#441)
Fix hanging the C-Vise process when a Pebble job never reports its timeout: the
logic to force-stop the job didn't actually take place if no other job was
executed, since we never returned from a wait() call.

Now we'll exit from wait() when it comes to the time when a job is considered as
a hung.